license-classifications.yml: Change license_set mapping
Rename the 'license_sets' section to 'categories', which is a less
technical term. Analogously, rename the LicenseSet class to
LicenseCategory.

Simplify categories by dropping the id property; as the id is an
arbitrary string, there is no point in having this field in addition
to the category name.

Categories are now an abstract concept. They can be freely defined by
the user, but need to be listed explicitly in the 'categories' section
of license-classifications.yml. LicenseConfiguration checks for
licenses referencing unknown categories.

Rename the 'licenses' section to 'categorizations', as its main
purpose is the assignment of licenses to categories.

These changes in the license model affect the rules file under
examples as well.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>