Replace flags in License by generic categories
Remove the flags in the License class controlling whether the license
should be included in the notice file and whether source code needs to
be provided. These flags are a bit arbitrary, and their meaning can be
expressed in a more flexible way by using categories.

While the 'include_source_code_offer_in_notice_file' has not been
used, the 'include_in_notice_file' property was accessed by
NoticeTemplateReporter allowing templates of the report generator the
filtering of licenses when generating notice files. Rework the class
and the templates to query for the corresponding category instead.

NoticeTemplateReporterFunTest did not check the correct evaluation of
the 'include_in_notice_file' property. Extend this test, so that this
property is now true for one license and false for another. Update the
expected results to reflect that only one license should now be
present in the resulting notice file.

Change the categories field from a SortedSet to a Set to enforce
immutability; the ordering of category names is not needed here.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>